refactor(sagas): rename file surveys to survey
