+++ src/reducers/index.js
@@ -1,5 +1,5 @@
 import network from './network';
-import surveys from './surveys';
+import surveys from './survey';
 import sync from './sync';
 
 export {network};
