+++ src/services/index.js
@@ -1,5 +1,5 @@
 import ping from './ping';
-import surveys from './surveys';
+import surveys from './survey';
 import sync from './sync';
 
 export {ping};
